pyproject.toml
[tool.poetry]
name = "cook-helper"
version = "0.1.0"
description = ""
authors = [
        "Artem Zlobenko <artemzlobenko@gmail.com>",
        
    ]
readme = "README.md"
packages = [{include = "cook_helper"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
requests = "^2.28.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
